refactor: flatten file structure
